Making printer as subscriber and configurable urls in cmd line
